sourcecode/joint_model/requirements.txt
alabaster==0.7.12
allennlp==0.8.4
apex==0.1
atomicwrites==1.3.0
attrs==19.1.0
awscli==1.16.198
Babel==2.7.0
bert-serving-client==1.8.9
bert-serving-server==1.9.3
blis==0.2.4
boto==2.49.0
boto3==1.9.188
botocore==1.12.188
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
conllu==0.11
cycler==0.10.0
cymem==2.0.2
Cython==0.29.12
decorator==4.4.0
docutils==0.14
editdistance==0.5.3
en-core-web-sm==2.1.0
fairseq==0.7.1
flaky==3.6.0
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.5.1
gensim==3.8.0
gevent==1.4.0
GMatch4py==0.2.5b0
GPUtil==1.4.0
greenlet==0.4.15
h5py==2.9.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonlines==1.2.0
jsonnet==0.13.0
jsonpickle==1.2
jsonschema==2.6.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==7.1.0
murmurhash==1.0.2
networkx==2.3
neuralcoref==4.0
nltk==3.4.4
numpy==1.16.4
numpydoc==0.9.1
overrides==1.9
packaging==19.0
pandas==0.24.2
parsimonious==0.8.1
Pillow==6.1.0
plac==0.9.6
pluggy==0.12.0
portalocker==1.4.0
preshed==2.0.1
protobuf==3.9.0
psutil==5.6.3
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.4.2
pyparsing==2.4.0
pytest==5.0.1
python-dateutil==2.8.0
pytorch-pretrained-bert==0.6.2
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.2
regex==2019.6.8
requests==2.22.0
responses==0.10.6
rsa==3.4.2
s3transfer==0.2.1
sacrebleu==1.3.7
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
smart-open==1.8.4
snowballstemmer==1.9.0
spacy==2.1.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
srsly==0.0.7
tensorboardX==1.8
termcolor==1.1.0
thinc==7.0.8
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.2
typing==3.7.4
Unidecode==1.1.1
urllib3==1.25.3
wasabi==0.2.2
wcwidth==0.1.7
Werkzeug==0.15.4
word2number==1.1
zipp==0.5.2
